requirements.txt
pyyaml
label-studio-sdk
